backend/requirements.txt
fastapi
uvicorn
openai
supabase-py
python-dotenv 